fix: rename folder and package name to make them the same
