+++ golang_api_server/pkg/data_access/elasticsearch_data_access/search_query.go
@@ -1,4 +1,4 @@
-package es_dao
+package elasticsearch_data_access
 
 type QueryBody struct {
 	From      int           `json:"from,omitempty"`
