+++ golang_api_server/pkg/data_access/elasticsearch_data_access/search_response.go
@@ -1,4 +1,4 @@
-package es_dao
+package elasticsearch_data_access
 
 import "encoding/json"
